[client] Seperated frontend from backend logic.
This commit takes out the logic from all UI code and moves it
into a dedicated backend service.
From now on, there should be no more protocol or program logic
in catmailclient.py or in one of the view classes.
The only logic remaining in catmailclient.py is used to control
the UI.